fixing folder structure and stripe node
