+++ client/src/pages/Home/Home.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import Wrapper from '../../components/Wrapper';
 import Header from '../../components/Header';
-import PillarSection from '../../components/PillarSection';
-import Mission from '../../components/Mission';
+import PillarSection from '../../components/HomeComponents/PillarSection';
+import Mission from '../../components/HomeComponents/Mission';
 import Footer from '../../components/Footer/Footer'; 
 import Jumbotron from '../../components/HomeComponents/Jumbotron';
